requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
pandas==2.2.2
openpyxl==3.1.2
numpy==1.26.4
scipy==1.13.0
firebase-admin==6.5.0
python-dotenv==1.0.1
